% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_rt_scatter_plots.R
\name{create_mz_rt_scatter_plots}
\alias{create_mz_rt_scatter_plots}
\title{create mz_rt scatter plot}
\usage{
create_mz_rt_scatter_plots(input_df)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
a ggplot scatter plot
}
\description{
This function loads a file as a dataframe.
}
