% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclomatic_complexity.R
\name{cyclomatic_complexity}
\alias{cyclomatic_complexity}
\title{Calculate the cyclomatic complexity}
\usage{
cyclomatic_complexity(file)
}
\arguments{
\item{file}{The file to calculate for - a file-path.}
}
\value{
The cyclomatic complexity for each function in a file.
}
\description{
Calculate the cyclomatic complexity for each function in a file
}
