% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolites_CV.R
\name{get_CV}
\alias{get_CV}
\title{A get_CV function}
\usage{
get_CV(pp_meta_df, df_name = "Vital")
}
\arguments{
\item{pp_meta_df}{input data frame}
}
\value{
a ggplot line graph
}
\description{
This function allows you to get metabolite CV.
}
