% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imputated_data.R
\name{get_imputated_data}
\alias{get_imputated_data}
\title{A get_imputated_data function}
\usage{
get_imputated_data(input_df)
}
\arguments{
\item{input_df, }{input data frame.}
}
\value{
a r dataframe without NA values
}
\description{
This function allows you to imputate the data, get rid of NA
}
