% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_vs_running_order.R
\name{get_median_vs_running_order}
\alias{get_median_vs_running_order}
\title{A median_vs_running_order function}
\usage{
get_median_vs_running_order(df_input)
}
\arguments{
\item{input_df, }{input data frame.}
}
\value{
a ggplot scatter graph
}
\description{
This function allows you to get median vs running order
}
