% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_vs_running_order_sample_vs_pp.R
\name{get_median_vs_running_order_comparision}
\alias{get_median_vs_running_order_comparision}
\title{A get_median_vs_running_order_comparision  function}
\usage{
get_median_vs_running_order_comparision(sample_meta_data, pp_meta_data,
  df_name = "Vital")
}
\arguments{
\item{sample_meta_data, }{input data frame.}

\item{pp_meta_data, }{input data frame.}

\item{df_name, }{string for data set}
}
\value{
a ggplot scatter graph
}
\description{
This function allows you to get median vs running order comparision between sample and pp
}
