% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SD_vs_Median.R
\name{get_median_vs_SD}
\alias{get_median_vs_SD}
\title{create get_median_vs_SD scatter plot}
\usage{
get_median_vs_SD(input_df)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
a ggplot scatter plot
}
\description{
This function loads a file as a dataframe, create scatter plot for median vs SD
}
