% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missingness_vs_running_order_comparision.R
\name{get_missingness_vs_running_order_comparision}
\alias{get_missingness_vs_running_order_comparision}
\title{A get_missingness_vs_running_order_comparision function}
\usage{
get_missingness_vs_running_order_comparision(sample_meta_data,
  pp_meta_data, df_name = "Vital")
}
\arguments{
\item{sample_meta_data, }{input data frame.}

\item{pp_meta_data, }{input data frame.}

\item{df_name, }{string for data set}
}
\value{
a ggplot line graph
}
\description{
This function allows you to get see the missingness of samples and pp
}
