% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plate_effect_by_PCA.R
\name{get_plate_effect_by_PCA}
\alias{get_plate_effect_by_PCA}
\title{A get_plate_effect_by_PCA function}
\usage{
get_plate_effect_by_PCA(input_df)
}
\arguments{
\item{input_df, }{input data frame.}
}
\value{
a r baseplot scatter plot graph
}
\description{
This function allows you to see plate effect using PCA.
}
