% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_waterfall_plots.R
\name{get_waterfall_plots}
\alias{get_waterfall_plots}
\title{A get_waterfall_plots function}
\usage{
get_waterfall_plots(input_df, clinical_df)
}
\arguments{
\item{input_df, }{input data frame.}
}
\value{
a r baseplot scatter plot graph
}
\description{
This function allows you to see waterfall plots for fishoil treatment.
}
