% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halstead.R
\name{halstead_measures}
\alias{halstead_measures}
\title{Calculate halstead software complexity measures}
\usage{
halstead_measures(file)
}
\arguments{
\item{file}{The file to calculate for - a file-path.}
}
\value{
A list with following elements:
\itemize{
\item{program vocabulary}
\item{program length}
\item{program length (estimated)}
\item{program volume}
\item{program difficulty}
\item{program effort}
\item{time to program (minutes)}
\item{number of bugs}
}
}
\description{
Calculate the halstead software complexity measures for a given file
}
\details{
See also \url{ https://en.wikipedia.org/wiki/Halstead_complexity_measures }
}
