% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC.R
\name{ICC}
\alias{ICC}
\title{An ICC function}
\usage{
ICC(meta, infor, is_full_set = TRUE, df_name = "Vital")
}
\arguments{
\item{infor, }{input data frame: metabolites infor}

\item{is_full_set, }{boolean}

\item{df_name, }{name for the input data}

\item{df_input, }{input data frame: metabolites data}
}
\value{
a r baseplot scatter plot graph
}
\description{
This function allows you to get ICC.
}
