% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_of_code.R
\name{lines_of_code}
\alias{lines_of_code}
\title{LOC calculator}
\usage{
lines_of_code(file)
}
\arguments{
\item{file}{The file to calculate for - a file-path.}
}
\value{
The list with two entries - the total and the non-empty lines of code.
}
\description{
Calculate the lines of code in a file
}
