% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_of_comments.R
\name{lines_of_comments}
\alias{lines_of_comments}
\title{Lines of comments for a file}
\usage{
lines_of_comments(file)
}
\arguments{
\item{file}{The file to calculate for - a file-path.}
}
\value{
The list with two entries - the documentation comments, and the actual
comments.
}
\description{
Calculate the total lines of comments for a file
}
