% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_SNE.R, R/t_SNE_for_metabolites_group.R
\name{make_t_SNE_graph_for_metabolites_group}
\alias{make_t_SNE_graph_for_metabolites_group}
\title{A get_CV function}
\usage{
make_t_SNE_graph_for_metabolites_group(input_df)

make_t_SNE_graph_for_metabolites_group(input_df)
}
\arguments{
\item{input_df, }{input data frame.}

\item{input_df, }{input data frame.}
}
\value{
a ggplot line graph

a r baseplot scatter plot graph
}
\description{
This function allows you to get metabolite CV.

This function allows you to see metabolites group.
}
