% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_SNE_for_plate_effects.R
\name{make_t_SNE_graph_for_plate_effects}
\alias{make_t_SNE_graph_for_plate_effects}
\title{A make_t_SNE_graph_for_plate_effects function}
\usage{
make_t_SNE_graph_for_plate_effects(df_input, is_sample = TRUE,
  df_name = "Vital")
}
\arguments{
\item{df_input, }{input data frame.}

\item{is_sample, }{boolean}

\item{df_name, }{name for the input data}
}
\value{
a r baseplot scatter plot graph
}
\description{
This function allows you to see plate effect.
}
