% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_info_data}
\alias{sample_info_data}
\title{metabolites information with plate_well as id}
\format{A data frame with columns:
\describe{
 \item{plate_well}{string in the "0001_0001 format, the last well for plate one will be "0001_0096}
 \item{subjectId}{ an interger}
 \item{year}{ 0 or 1, an interger}
}}
\source{
modified from https://stattarget.github.io/docs/demo/
}
\usage{
sample_info_data
}
\description{
metabolites information with plate_well as id
}
\examples{
\dontrun{
 sample_info_data
}
}
\keyword{datasets}
