% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_meta_data}
\alias{sample_meta_data}
\title{metabolites reading with plate_well as id}
\format{A data frame with columns:
\describe{
 \item{plate_well}{string in the "0001_0001 format, the last well for plate one will be "0001_0096}
 \item{Eico_mzid_383.182643_4.2555}{ a metabolites produced from the experiment, it is a combination of group, "mzid", mz and retension time}
}}
\source{
modified from https://stattarget.github.io/docs/demo/
}
\usage{
sample_meta_data
}
\description{
metabolites reading with plate_well as id
}
\examples{
\dontrun{
 sample_meta_data
}
}
\keyword{datasets}
