% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_trt_data}
\alias{sample_trt_data}
\title{metabolites information with subjectId and year as id}
\format{A data frame with columns:
\describe{
 \item{subjectId}{an interger}
 \item{year}{ 0 or 1, an interger}
 \item{fishoilactive}{ 0 or 1, an interger}
 \item{vitdactive}{ 0 or 1, an interger}
}}
\source{
radomn generated
}
\usage{
sample_trt_data
}
\description{
metabolites information with subjectId and year as id
}
\examples{
\dontrun{
 sample_trt_data
}
}
\keyword{datasets}
