% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/software_maintainability.R
\name{software_maintainability}
\alias{software_maintainability}
\title{Calculate software maintainability}
\usage{
software_maintainability(file)
}
\arguments{
\item{file}{The file to calculate for - a file-path.}
}
\value{
The maintainability index (a number).
}
\description{
Calculate the software maintainability index for a given file
}
