% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_type_ratio.R
\name{token_type_ratio}
\alias{token_type_ratio}
\title{Token type ratio}
\usage{
token_type_ratio(file)
}
\arguments{
\item{file}{The file to calculate for - a file-path.}
}
\value{
The ratio of tokens present in the file (normalised so all tokens
together sum to 1).
}
\description{
Calculate the ratio of tokens present in a file
}
