% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_per_line.R
\name{tokens_per_line}
\alias{tokens_per_line}
\title{Calculate tokens pers line}
\usage{
tokens_per_line(file)
}
\arguments{
\item{file}{The file to calculate for - a file-path.}
}
\value{
The number of tokens per line for the entire file.
}
\description{
Calculate the number of tokens per line for a file
}
