% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt_logit_model.R
\name{trt_logit_model}
\alias{trt_logit_model}
\title{A trt_logit_model  function}
\usage{
trt_logit_model(sample_delta_data, sample_trt_data,
  trt = "fishoilactive", df_name = "Vital")
}
\arguments{
\item{sample_trt_data, }{input data frame with treatment info.}

\item{trt, }{string showing trtment, which is a var i nthe sample trt data}

\item{df_name, }{string for data set}

\item{sample_meta_data, }{input data frame with metabolites reading.}
}
\value{
a cvs files with beta and p value for treatment effects
}
\description{
This function allows you to get median vs running order comparision between sample and pp
}
