% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt_logit_model_with_age_sex_gender.R
\name{trt_logit_model_with_age_gender}
\alias{trt_logit_model_with_age_gender}
\title{A trt_logit_model_with_age_gender  function}
\usage{
trt_logit_model_with_age_gender(sample_delta_data, sample_trt_data,
  sample_clin_data, trt = "fishoilactive", df_name = "Vital")
}
\arguments{
\item{sample_trt_data, }{input data frame with trt info.}

\item{sample_clin_data, }{input data frame with clinical info.}

\item{trt, }{string showing trtment, which is a var i nthe sample trt data}

\item{df_name, }{string for data set}

\item{sample_meta_data, }{input data frame with metabolites reading.}
}
\value{
a cvs files with beta and p value for treatment effects
}
\description{
This function allows you to get median vs running order comparision between sample and pp
}
