% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_rt_scatter_plots.R
\name{create_mz_rt_scatter_plots}
\alias{create_mz_rt_scatter_plots}
\title{create mz_rt scatter plot}
\usage{
create_mz_rt_scatter_plots(input_df, df_name = "Vital")
}
\arguments{
\item{input_df, }{input data frame.}

\item{df_name, }{name for the input data, default is "Vital"}

\item{infile}{Path to the input file}
}
\value{
a ggplot scatter plot
}
\description{
This function loads a file as a dataframe.
}
