% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolites_CV.R
\name{get_CV}
\alias{get_CV}
\title{A get_CV function}
\usage{
get_CV(sample_meta_df, sample_info_df, df_name = "Vital")
}
\arguments{
\item{sample_meta_df, }{sample_info_df}
}
\value{
a ggplot line graph
}
\description{
This function allows you to get metabolite CV.
}
