% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missingness_vs_running_order.R
\name{get_missingness_vs_running_order}
\alias{get_missingness_vs_running_order}
\title{A get_missingness_vs_running_order function}
\usage{
get_missingness_vs_running_order(input_df, df_name)
}
\arguments{
\item{input_df, }{input data frame.}
}
\value{
a ggplot line graph
}
\description{
This function allows you to get see the missingness of samples and pp
}
