% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missingness_vs_running_order.R,
%   R/get_missingness_vs_running_order_comparision.R
\name{get_missingness_vs_running_order}
\alias{get_missingness_vs_running_order}
\title{A get_missingness_vs_running_order function}
\usage{
get_missingness_vs_running_order(sample_meta_data, pp_meta_data,
  df_name = "Vital")

get_missingness_vs_running_order(sample_meta_data, pp_meta_data,
  df_name = "Vital")
}
\arguments{
\item{sample_meta_data, }{pp_meta_data, input data frame.}

\item{meta_data, }{input data frame.}
}
\value{
a ggplot line graph

a ggplot line graph
}
\description{
This function allows you to get see the missingness of samples and pp

This function allows you to get see the missingness of samples and pp
}
