% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_SNE_for_plate_effects.R
\name{make_t_SNE_graph_for_plate_effects}
\alias{make_t_SNE_graph_for_plate_effects}
\title{A make_t_SNE_graph_for_plate_effects function}
\usage{
make_t_SNE_graph_for_plate_effects(input_df, df_name)
}
\arguments{
\item{input_df, }{input data frame.}
}
\value{
a r baseplot scatter plot graph
}
\description{
This function allows you to see plate effect.
}
