% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt_logit_model.R
\name{trt_logit_model}
\alias{trt_logit_model}
\title{A trt_logit_model  function}
\usage{
trt_logit_model(sample_meta_data, sample_trt_data, df_name = "Vital")
}
\arguments{
\item{sample_meta_data, }{input data frame.}

\item{df_name, }{string for data set}

\item{pp_meta_data, }{input data frame.}
}
\value{
a ggplot scatter graph
}
\description{
This function allows you to get median vs running order comparision between sample and pp
}
